fix(refactor): useOnClickOutside 통합 및 버그 수정
- 어딜 클릭해도 아래로 내려가던 버그
