+++ src/components/chart/IndicatorToolbar.tsx
@@ -4,7 +4,7 @@ import { useMemo, useRef, useState } from 'react';
 import { createPortal } from 'react-dom';
 import { cn } from '@/lib/cn';
 import { getPeriodColor } from '@/lib/chartColors';
-import { useOnClickOutside } from '@/components/chart/hooks/useOnClickOutside';
+import { useOnClickOutside } from '@/components/hooks/useOnClickOutside';
 
 interface IndicatorToggleGroup {
     visible: boolean;
@@ -161,7 +161,7 @@ interface IndicatorToolbarProps {
     cci: IndicatorToggleGroup;
     volumeProfile: IndicatorToggleGroup;
     ichimoku: IndicatorToggleGroup;
-    candlePatterns?: IndicatorToggleGroup;
+    _candlePatterns?: IndicatorToggleGroup;
 }
 
 export function IndicatorToolbar({
@@ -180,7 +180,7 @@ export function IndicatorToolbar({
     cci,
     volumeProfile,
     ichimoku,
-    candlePatterns,
+    _candlePatterns,
 }: IndicatorToolbarProps) {
     const [isExpanded, setIsExpanded] = useState(false);
     const [openDropdown, setOpenDropdown] = useState<DropdownType>(null);
